Added handleLogin to api.ts, fixed navigate bug.
